Implemented addition/subtraction in Assembly
~subtraction is not yet fully implemented
+minor fixes
+pseudo random generator and unit tests (kind of)

+++ Message.cpp
@@ -22,7 +22,7 @@ void Message::embedMessage(const string &message)
 {
 	if(length != 0)
 		delete [] messageData;
-	unsigned messageSize = message.size();
+	size_t messageSize = message.size();
 	int bytesInChunk = chunkSize*(BITS_IN_CHUNK/8);
 	length = messageSize/(bytesInChunk-1)+1;
 	messageData = new mpuint[length];
@@ -80,7 +80,7 @@ string Message::extractMessage()
 
 void Message::encryptMessage(const mpuint &e,const mpuint &n)
 {
-	for(int i=0;i<length;i++)
+	for(unsigned i=0;i<length;i++)
 	{
 		mpuint result = mpuint(messageData[i].length);
 		mpuint::Power(messageData[i],e,n,result);
@@ -90,7 +90,7 @@ void Message::encryptMessage(const mpuint &e,const mpuint &n)
 
 void Message::decryptMessage(const mpuint &d,const mpuint &n)
 {
-	for(int i=0;i<length;i++)
+	for(unsigned i=0;i<length;i++)
 	{
 		mpuint result = mpuint(messageData[i].length);
 		mpuint::Power(messageData[i],d,n,result);
